Files for project topic: CBAM
Data are provided in different file formats, but all contain the same information. Some general information on the data set and variables are summarized in the excel file: data.xls

The data reports information on the EU27's largest 100 trade partners (in terms of their share in total merchandise imports by the EU), the sectoral structure of their trade with the EU and the rest of the world (RoW), the average CO2 emission intensity of these sectors, as well as the countries' total GDP and share in world merchandise trade.