Files for project topic: Political Economy of the Trade War
Data are provided in different file formats, but all contain the same information. Some general information on the data set and variables are summarized in the excel file: data.xls

The data reports information on reecent presidential election outcomes at US  state levels, the sectoral employment structure, as well as information on US trade and tariffs with China and the rest of the world.