Files for project topic: Japan vs. South Korea
Data are provided in different file formats, but all contain the same information. Some general information on the data set and variables are summarized in the respective excel files [xls]

The data report information on South Korea's imports of three critical production inputs, both at annual and monthly frequency and for different trade partners. There is also a data set with information on South Korea's exports of various semiconductors (for which the imported goods serve as critical inputs).