Files for project topic: BRI Africa
Data are provided in different file formats, but all contain the same information. Some general information on the data set and variables are summarized in the excel file: data.xls

The data reports information on annual import values of China from African countries, their GDP and the amount of Chinese investment in these economies (BRI or non-BRI, in value and numbers).