Files for project topic: Political Economy of the Trade war
Data are provided in different file formats. The Excel file is more aggregated and ready for use. The other files (csv and dta) are more disaggregated and allow for alternative calculation methods. Some general information on the data set and variables are summarized in the excel file: data.xls

The data reports information on US county-level outcomes in the 2016 and 2020 US presidential elections. There is additional information on counties sectoral employment structure prior to the elections, as well as US-wide industry level exposure to US tariffs on Chinese imports, retaliatory tariffs by China on US exports, and respective import and export data. That information can be used to compute counties exposure to the tariffs.