AEI - Chinese Investment Tracker
This dataset comprises a subset of the AEI investment tracker dataset, reporting Chinese investments in African countries, by sector, total number and value of projects per year. The period covered is 2005-2021. The compiled data is based on the Fall-2023 vintage of this dataset. The American Enterprise Institute regularly updates this dataset, which is publicly available on its website: https://www.aei.org/china-global-investment-tracker/